+++ src/cli/lib/styles/font.ts
@@ -22,9 +22,17 @@ const fontWeightVariants = {
 export const font: Parser = {parse: fontParser};
 
 function fontParser(src: Font) {
+  const cssArr: string[] = [''];
+  const fontFamilies: string[] = [];
   const fontStyle = fontWeightVariants[src.style];
 
-  return `;font-family:${src.family}
-  ;font-size:${src.size}px
-  ;font-weight:${fontStyle}`;
+  if (src.postscriptName) fontFamilies.push(src.postscriptName);
+  if (src.family) fontFamilies.push(src.family);
+  if (fontFamilies.length) cssArr.push(`font-family: ${fontFamilies.join(', ')}`);
+  if (fontStyle) cssArr.push(`font-weight: ${fontStyle}`);
+  if (src.size) cssArr.push(`font-size: ${src.size}px`);
+
+  cssArr.push('');
+
+  return cssArr.join(';');
 }